Remove pointless noise

+++ server/asobj/Date.cpp
@@ -100,7 +100,7 @@
 #endif
 
 #ifndef NAN
-	#define NAN static_cast<const float>(0.0/0.0)
+	#define NAN (0.0/0.0)
 #endif
 
 // Declaration for replacement timezone functions
